+++ src/predicates/constructions/mod.rs
@@ -1,0 +1,3 @@
+mod line_intersection;
+
+pub(crate) use line_intersection::line_intersection;